feat: Refactor getSectionTitle function and optimize section title retrieval in alerts and case management pages
